Break out style helpers into seperate file
